+++ internal/wasm/jit/jit_other.go
@@ -18,3 +18,15 @@ func jitcall(codeSegment, ce uintptr) {
 func newCompiler(f *wasm.FunctionInstance, ir *wazeroir.CompilationResult) (compiler, func(), error) {
 	panic("unsupported GOARCH")
 }
+
+const (
+	reservedRegisterForCallEngine              = 0
+	reservedRegisterForStackBasePointerAddress = 0
+	reservedRegisterForMemory                  = 0
+	reservedRegisterForTemporary               = 9
+)
+
+var (
+	generalPurposeFloatRegisters         = []int16{}
+	unreservedGeneralPurposeIntRegisters = []int16{}
+)
