+++ internal/wasm/jit/jit_value_location.go
@@ -1,5 +1,3 @@
-//go:build amd64 || arm64
-
 package jit
 
 import (
